refactor(All) : 패키지 폴더명 변경 , 테스트 내 패키지 폴더 생성
누락된 패키지 명 reward -> result 변경
테스트 코드에도 폴더 만들어서 도메인 테스트 간 분리
