+++ src/main/java/controller/GameController.java
@@ -7,8 +7,8 @@ import domain.common.Name;
 import domain.player.Names;
 import domain.player.Player;
 import domain.player.Players;
-import domain.reward.PlayerResult;
-import domain.reward.Results;
+import domain.result.PlayerResult;
+import domain.result.Results;
 import util.RandomDirectionGenerator;
 import view.InputView;
 import view.LadderCommand;
