+++ src/main/java/domain/result/PlayerResult.java
@@ -1,4 +1,4 @@
-package domain.reward;
+package domain.result;
 
 import domain.common.Name;
