+++ src/test/java/domain/result/ResultTest.java
@@ -1,6 +1,6 @@
-package domain;
+package domain.result;
 
-import domain.reward.Result;
+import domain.result.Result;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
